% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-proj.R
\name{convert_proj}
\alias{convert_proj}
\title{Convert Proj}
\usage{
convert_proj(data, data.x = "Long", data.y = "Lat",
  data.CRS = "+init=epsg:4326", new.CRS = "+init=epsg:3005")
}
\arguments{
\item{data}{The data.frame to convert.}

\item{data.x}{A string of the name of the column containing the longitude.}

\item{data.y}{xx}

\item{data.CRS}{xx}

\item{new.CRS}{xx}
}
\value{
A SpatialPointsDataFrame with
}
\description{
Convert Proj
}

