% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cumshewa_wind}
\alias{cumshewa_wind}
\title{Cumshewa Wind}
\format{A tbl data frame:
\describe{
  \item{Station}{The station name (chr).}
  \item{DateTime}{The date and time in Pacific Standard Time (UTC - 8) (dttm).}
  \item{Direction}{The direction the wind is blowing from in degrees (dbl).}
  \item{Speed}{The wind speed in km/hour (dbl).}
  \item{Long}{The station's longitude in EPSG:4326 (dbl).}
  \item{Lat}{The station's latitude in EPSG:4326 (dbl).}
}}
\usage{
cumshewa_wind
}
\description{
A tbl of the hourly wind direction and speed at Cumshewa Head 2012 to 2016.
}
\details{
Hourly wind measurements from \url{http://http://climate.weather.gc.ca/historical_data/search_historic_data_e.html}.
}
\keyword{datasets}
