% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg-fetch.R
\name{hg_fetch}
\alias{hg_fetch}
\title{Haida Gwaii Fetch}
\usage{
hg_fetch()
}
\value{
A tbl with the columns Station, DateTime, Direction, Speed, Long and Lat.
}
\description{
Downloads and binds the eight fetch datasets from \url{https://raw.githubusercontent.com/sebdalgarno/haidawave-data/master/}.
}
\details{
Points were generated within a GIS spaced 10m along 1:50,000 ShoreZone vector coastline data (Howes et al. 1994).
From each point, fetch distance (i.e. distance from point to nearest ShoreZone coastline) was calculated at 5 degree intervals using
the java-based software - 'Vector_fetch' - developed by Mika Murtojarvi (Murtojarvi et al. 2007).
Since ShoreZone vector coastline data were limited to BC, fetch distances are only accurate to 200,000 m.

# ## References
}
