% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg-wind.R
\name{hg_wind}
\alias{hg_wind}
\title{Haida Gwaii Wind}
\usage{
hg_wind()
}
\value{
A tbl with the columns Station, DateTime, Direction, Speed, Long and Lat.
}
\description{
Downloads the dataset from \url{https://raw.githubusercontent.com/sebdalgarno/haidawave-data/master/wind-hg-1216.csv}.
}
\details{
The wind data set comprises hourly wind observations for all weather stations in Haida Gwaii with available data from 2012-2016.
These data were taken from <http://climate.weather.gc.ca/historical_data/search_historic_data_e.html>.
}
