% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-fetch.R
\name{mean_fetch}
\alias{mean_fetch}
\title{Calculate fetch.}
\usage{
mean_fetch(x, max_fetch = 2 * 10^5)
}
\arguments{
<<<<<<< HEAD
\item{site.data}{A SpatialPoints object containing sites that require estimates of wave exposure.}

\item{fetch.data}{A SpatialPointsDataFrame containing fetch distance data.}

\item{max.distance}{An integer indicating desired maximum fetch distance (in metres).}
=======
\item{x}{A data frame with the column Fetch.}

\item{max_fetch}{A number indicating the maximum fetch distance to consider.}
>>>>>>> dev
}
\value{
A data frame with a single row and column Fetch indicating the mean fetch.
}
\description{
Calculates mean fetch.
}
\examples{
mean_fetch(haidawave::laskeek_fetch)
}
