% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest-point.R
\name{nearest_point}
\alias{nearest_point}
\title{Find nearest point.}
\usage{
nearest_point(data1, data2)
}
\arguments{
\item{data1}{A SpatialPointsDataFrame.}

\item{data2}{A SpatialPointsDataFrame.}
}
\value{
One additional column added to data1 indicating distance to nearest point (in units of projection): 'NearestDistance'.
}
\description{
Find the nearest point and ID from one SpatialPointsDataFrame to another.
}
\details{
Note that the convert_proj function can be used first to convert data to SpatialPointsDataFrame and ensure identical CRS.
}

