% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum-fetch.R
\name{sum_fetch}
\alias{sum_fetch}
\title{Calculate summed Fetchs.}
\usage{
sum_fetch(site.data, fetch.data, fetch.ID = "PointID",
  fetch.distance = "Distance", max.distance = 2e+05)
}
\arguments{
\item{site.data}{A SpatialPointsDataFrame of sites requiring estimates of wave exposure.}

\item{fetch.data}{A SpatialPointsDataFrame with columns containing unique ID, bearing, and fetch distance, along with coordinates.}

\item{fetch.ID}{A string of the name of the column containing a unique point ID.}

\item{fetch.distance}{A string of the name of the column containing fetch distance.}

\item{max.distance}{An integer indicating desired maximum dfetch distance.}
}
\value{
Two additional columns to site.data: 'nearest.pt' contains unique ID of nearest fetch.data point; 'sumfetch' contains summed fetch results.
}
\description{
Set maximum fetch distance and calculate summed fetch from nearest point.
}
\details{
For this function to work, input data must be SpatialPointsDataFrame with identical CRS. The convert_proj function can be used to achieve that. Note that it is not necessary to run the nearest_point function prior to this one.
To reduce processing time, subset fetch.data to a region encompassing your site.data.
}

