% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-fetch.R
\name{weighted_fetch}
\alias{weighted_fetch}
\title{Calculate weighted mean fetch.}
\usage{
weighted_fetch(x, max_fetch = 2 * 10^5)
}
\arguments{
\item{x}{A data frame with the columns Fetch and Weight.}

\item{max_fetch}{A number indicating the maximum fetch distance to consider.}
}
\value{
A data frame with a single row and column Fetch indicating the mean fetch.
}
\description{
Calculates weighted mean fetch.
}
