% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind-weights.R
\name{wind_weights}
\alias{wind_weights}
\title{Calculate wind weighting for each provided direction.}
\usage{
wind_weights(x)
}
\arguments{
\item{x}{A data frame with the columns Direction and Speed.}
}
\value{
A tbl with the columns Direction, Frequency, Speed and Weight.
}
\description{
Calculates weighting based on wind direction and speed.
It takes a data frame and returns a tbl with the columns Direction,
Frequency (number of intervals spent blowing in that direction),
Speed (the average speed when blowing in that direction) and
Weight ()
}
\examples{
wind_weights(haidawave::cumshewa_wind)
}
