% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-proj.R
\name{convert_proj}
\alias{convert_proj}
\title{Convert projection.}
\usage{
convert_proj(x, data.CRS = "+init=epsg:4326", new.CRS = "+init=epsg:3005")
}
\arguments{
\item{x}{The data.frame to convert.}

\item{data.CRS}{PROJ4 string defining current CRS of data.}

\item{new.CRS}{PROJ4 string of desired CRS.}
}
\value{
A tbl with X and Y transformed to desired CRS.
}
\description{
Convert a data.frame to a SpatialPointsDataFrame with desired CRS.
The data.frame must contains the columns X and Y for the x and y coordinates.
}
\details{
The default is to convert WGS84 Longitude/Latitude to BC Albers.
}
\examples{
convert_proj(data.frame(Site = 1, X = c(-131.504), Y = c(52.871)))
}
