% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cumshewa_wind}
\alias{cumshewa_wind}
\title{Cumshewa Wind}
\format{A tbl data frame:
\describe{
  \item{Station}{The station name (chr).}
  \item{DateTime}{The date time (time).}
  \item{MLLW}{The tide height in m (dbl).}
}}
\usage{
cumshewa_wind
}
\description{
A tbl of the hourly wind direction and speed at Cumshewa Head 2012 to 2016.
}
\details{
High/Low Tide Predictions from \url{http://tidesandcurrents.noaa.gov/tide_predictions.html}.
}
\keyword{datasets}

