% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cumshewa_wind}
\alias{cumshewa_wind}
\title{Cumshewa Wind}
\format{A tbl data frame:
\describe{
  \item{Station}{The station name (chr).}
  \item{DateTime}{The date time (time).}
  \item{Direction}{The wind direction in degrees (int).}
  \item{Speed}{The wind speed in km/hour (int).}
  \item{Long}{The longitude of weather station, EPSG:4326 (dbl).}
  \item{Lat}{The latitude of weather station, EPSG:4326 (dbl).}
}}
\usage{
cumshewa_wind
}
\description{
A tbl of the hourly wind direction and speed at Cumshewa Head 2012 to 2016.
}
\details{
Hourly wind measurements from \url{http://http://climate.weather.gc.ca/historical_data/search_historic_data_e.html}.
}
\keyword{datasets}
