% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-fetch.R
\name{mean_fetch}
\alias{mean_fetch}
\title{Calculate mean fetch at sites.}
\usage{
mean_fetch(site.data, fetch.data, max.distance = 2e+05)
}
\arguments{
\item{site.data}{A SpatialPointsDataFrame of sites requiring estimates of wave exposure.}

\item{fetch.data}{A SpatialPointsDataFrame with fetch distance data.}

\item{max.distance}{An integer indicating desired maximum fetch distance (in metres).}
}
\value{
One additional column (MeanFetch) to site.data with mean fetch distance of nearest fetch point.
}
\description{
Set maximum fetch distance and assign mean fetch from nearest point.
}
\details{
Input data must be SpatialPoints with identical CRS. The convert_proj function can be used to achieve that.
Note that it is not necessary to run the nearest_point function first.
}

