% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest-point.R
\name{nearest_point}
\alias{nearest_point}
\title{Find nearest point.}
\usage{
nearest_point(data1, data2, data2.ID = "PointID")
}
\arguments{
\item{data1}{A SpatialPointsDataFrame.}

\item{data2}{A SpatialPointsDataFrame.}

\item{data2.ID}{A string of the name of the column in data2 with ID. If IDs are not unique, function will remove duplicates.}
}
\value{
One additional column added to data1: 'nearest.ID'.
}
\description{
Find the nearest point and ID from one SpatialPointsDataFrame to another.
}
\details{
Note that the convert_proj function can be used first to convert data to SpatialPointsDataFrame and ensure identical CRS.
}

