% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest-point.R
\name{nearest_point}
\alias{nearest_point}
\title{Joins y to x and y based on their X and Y coordinates.}
\usage{
nearest_point(x, y, suffix = c(".x", ".y"))
}
\arguments{
<<<<<<< HEAD
\item{data1}{Object of class SpatialPoints or SpatialPointsDataFrame.}

\item{data2}{Object of class SpatialPoints or SpatialPointsDataFrame.}
=======
\item{x}{A data frame with the columns X and Y.}

\item{y}{A data frame with the columns X and Y.}

\item{suffix}{If there are non-joined duplicate variables in \code{x} and
\code{y}, these suffixes will be added to the output to diambiguate them.}
>>>>>>> dev
}
\value{
The joined data frame with the additional column Distance indicating the distance between the X and Y coordinates.
}
\description{
<<<<<<< HEAD
For a given set of points, find the distance to the nearest point of another set.
=======
Joins x to y based on minimising the euclidean distance between their X and Y coordinates.
As a result each row in x is assigned the closest row in y.
>>>>>>> dev
}
\details{
Note the convert_proj function can be used to ensure the X and Y coordinates are in the same projection.
}
