% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind-fetch.R
\name{wind_fetch}
\alias{wind_fetch}
\title{Calculate wind-altered summed fetch.}
\usage{
wind_fetch(site.data, fetch.data, fetch.bearing = "Bearing",
  fetch.distance = "Distance", fetch.ID = "PointID", weights.data,
  weights.station = "Station", weights.direction = "Direction",
  weights = "Weight", max.distance = 2e+05)
}
\arguments{
\item{site.data}{A SpatialPointsDataFrame of sites requiring estimates of wave exposure.}

\item{fetch.data}{A SpatialPointsDataFrame of sites with a column indicating fetch bearing and a column indicating fetch distance.}

\item{fetch.bearing}{A string of the name of the column containing fetch bearing.}

\item{fetch.distance}{A string of the name of the column containing fetch distance.}

\item{fetch.ID}{A string of the name of the column containing unique ID.}

\item{weights.data}{A SpatialPointsDataFrame including columns indicating station name, direction and weighting factor.}

\item{weights.station}{A string of the name of the column containing names of weather stations.}

\item{weights.direction}{A string of the name of the column containing wind direction.}

\item{weights}{A string of the name of the column containing wind weighting factors for each direction.}

\item{max.distance}{An integer indicating desired maximum distance of fetch calculations.}
}
\value{
One additional column to site.data ('windfetch') with wind-altered summed fetch results.
}
\description{
Set maximum fetch distance, find fetch lines from nearest point, weight fetch lines according to conditions at nearest weather station, calculate wind-altered summed fetch.
}
\details{
For this function to work, input data must be SpatialPointsDataFrame with identical CRS. The convert_proj function can be used to achieve that.
Weighting factors (weights.data) should be generated first using the wind_weights function. Note that it is not necessary to run the nearest_point function prior to this one.
To reduce processing time, subset fetch.data to a region encompassing your site.data.
}

